Improve logic of retrieving language variant from Uri.
The previous implementation of the method for extracting the language
variant had some potential erroneous behavior, by assuming that the Uri
path must begin with either "/wiki/" or "/lang-variant/", which might not
always be the case.

This patch makes the method more fault-tolerant by explicitly checking for
more than one path separator, which is a more reliable indicator that the
first part of the path is either "/wiki/" or "/lang-variant/".

Change-Id: If08a5f9451ef993221c8ee3088a2a43a9ee91339