Make Offline Library enableable via Settings.
There is now a new Setting called "Enable Offline Library" that acts as
the gateway to enabling the entire Offline feature.

When this setting is enabled, it automatically triggers the dynamic
permission request that asks the user for read/write permissions, and goes
on to search the device for ZIM files.

This means that after enabling this new setting, the user doesn't actually
have to navigate to the Offline Library screen to start using their ZIM
files. They'll just start working.

Bug: T179072
Change-Id: Iecf652a9a12399cb8e8bbc3988776af942bc00cf