Slight optimization when parsing HTML from Strings.
There are very many places in our code where we call Html.fromHtml(),
because some of the strings that we receive from the server might contain
some minimal HTML formatting, such as bold, italics, and links.

However, if you dig into the code of the fromHtml() implementation, it's
actually pretty expensive and resource-intensive, and might actually
contribute to performance degradations when used too much.

The other thing is that *most* of the time, the strings we get from the
server don't actually contain any HTML formatting, so we're actually
passing the string into fromHtml() needlessly.

This patch makes it conditional whether or not a string will be parsed by
fromHtml(), based on whether the string contains at least one opening
bracket of an HTML tag ("<"). Otherwise, the string is returned as-is.

Change-Id: Ib57d6ede66e7ca3fec6d8b0b28b01db437007276