Slight performance improvement to Feed scrolling.
Most of our Feed cards have a "header" and a "footer", and previously
these components were dynamically inflated when the Feed card was bound to
the RecyclerView, instead of being inflated all at once when the view is
first created. This can actually cause a performance degradation in
certain cases.

This patch makes the header and footer views become part of the actual xml
layout of the card views, and will therefore inflate along with the rest
of the layout.

Change-Id: I4d77eafb4df381814a6c68f89cea1e62594a5abd