Introducing... ConstraintLayout!
This introduces the use of ConstraintLayout into many of our Feed
components. ConstraintLayout can greatly improve the performance of our
layouts because it makes it easy to flatten layouts that were previously
contained within nested Linear and FrameLayouts.

This will represent a further incremental improvement to the performance
of the Feed fragment as a whole.

In subsequent patches, we should try to apply ConstraintLayout wherever
applicable.

Change-Id: I5e5c2ebafcfdd765b03418fee7b5dc3ce4ab0b3c