No longer use 'noimages' API parameter.
This parameter had good intentions behind it, but ultimately it makes the
server do things that should really be done by the client.
...Especially since it can apparently also cause unexpected presentational
issues.
...And especially since the usage rate of 'noimages' is so low, that the
server shouldn't really need to worry about it.

This patch moves the responsibility of deciding whether or not to load
images entirely to the client.  When image loading is disabled, the
WebView will show a checkerboard image placeholder, like we currently do
in the feed.

Note: If the user explicitly clicks an image in the WebView, the Gallery
will be loaded with the correct full-size version of the image.

Bug: T137351
Bug: T103675
Change-Id: I2d19b1dec9e52ce252412c55271d1b6533f1df5d