Respect the full URL in gallery image description links
Our gallery link handling always constructs a page title based on the
current app language, but gallery image descriptions come from Wikimedia
Commons and may contain links to any wiki.

This updates the link handling code so as not to assume the app language
is the wiki being linked to, when a full URL is provided.

Bug: T114025
Change-Id: I8aa14f7a13d37a5f59edb397aa6dd11718a3c508