Use displaytitle from server for link previews
Prevents article title display inconsistencies by using the displaytitle
provided by the server response (and deferring adding the title to the
link preview until then). In practice, the delay should be barely user-
noticeable.

If the summary request fails, the app falls back to the old behavior in
the error callback.

This should also have the desirable effect of showing the true article
title in previews, with any redirects resolved.

Bug: T149429
Change-Id: I9030f5871250505e0e0399c94f2b857c30204db5