Black theme.
This adds a new Black theme to the app (mostly inherited from the Dark
theme), and also adds a new Activity for the user to "try on" any of the
themes that we offer.

This also provides image dimming for larger native images shown in the app
(i.e. images shown with FaceAndColorDetectImageView)

Bug: T173319
Change-Id: Ie885e6b8ea6014a93d9cc2b9874fc3622831f8d5