Fix interception of WebView content with extended content-type header.
Background:
We use a WebView to display article content, but we intercept the
WebView's network requests by using a custom WebViewClient where we
perform our own network transaction using OkHttp.

Once we receive a response, we pass it back to the WebView by creating a
WebResourceResponse object which contains the stream of bytes, as well as
the encoding and mime-type of the bytes.

The thing is:
The WebResourceResponse object expects the mime-type parameter to be a
pure mime-type string, that is "type/subtype", with nothing else appended
to it.  Whereas, we were incorrectly passing the entire "content-type"
header, which contains the mime-type, plus additional possible parameters
such as "charset" and "profile."

This patch updates our code to extract only the mime-type from the
content-type header, and correctly pass it into the WebResourceResponse,
thereby fixing the issues that users are seeing with SVG math formulas,
which were recently updated to contain a "charset" parameter in their
content-type.

Bug: T184347
Change-Id: I7e5417519e842995bb6a54a0ed6b22096827b67f