Extend HttpStatusException to contain ServiceError information.
This is a groundwork patch broken off from the big Syncing patch, to make
the changes more digestible.

The ServiceError structure gets transmitted along with HTTP 40x errors,
but we just weren't using it properly yet. This patch captures this
information, since we'll need it to differentiate between various service
errors when synchronizing.

Change-Id: Iee4d34d3bdd83ca6794ee6c79784b6dedbc2a8fa