Lazily load the Nearby fragment.
When the user goes to the History tab, the Nearby tab is pre-loaded
implicitly by the ViewPager, which causes much of the MapView logic to get
loaded unnecessarily. The MapView is a very heavy component, and
contributes to jitter and slowness when navigating between our bottom
tabs.

This patch encapsulates NearbyFragment inside a lazy-loadable fragment,
which is just a skeleton fragment that does not consume any resources when
loaded implicitly by the ViewPager.

Change-Id: Ic02db1b2bc5b7b27ca34b2f5034610c9a1c99476