Fix deep links for non /wiki/ links
In df4fb8073c0e88875d93d452bb3859b3f396a26d an intent-filter was modified to
handle URLs with a path prefix of the form `/zh-.*/`. However, this filter
matches more than it is supposed to.  The `path-prefix` and `path-pattern` are
not "prefix + pattern" but the filter matches if any of the two matches, so it
would actually match paths of the form `/zh.*` and `-.*/`. The latter is
impossible as it does not start with a slash.  The former includes paths such
as `/zhfoo` which should probably not be matched. Putting everything into a
`path-pattern` fixes this. Note that the two `<data>` entries are merged, as
explained at
https://developer.android.com/training/app-links/deep-linking.html, so
`http://` and `https://` applies to both.

Change-Id: I8a5b4e01bd1886fc392d307fa87f04dab3fa3f26