Enforce Unicode NFC when exchanging strings with Reading List service.
Since we'll be sharing the Reading List service with different platforms,
we need to be consistent about the precise encoding and composition of the
Unicode strings that represent the list names and article titles stored in
them.

This patch enforces Unicode Normalization Form C on the names and
descriptions of reading list names and article titles when they're
communicated back and forth with the Reading List service.

Change-Id: Ie4a89cb8a6cc7de1eebe35080c7556615564f8fa