Add more checks to protect and bullet-proof the Default list.
This adds several logical checks to prevent renaming or deleting the
Default list, both at the database level and at the UI level.

There can be a few cases where the UI might lead the user to allow
renaming or deleting the default list (e.g. through a RecyclerView quirk,
or through some other logic that we missed), and we need to prevent this
at all costs.

Bug: T185330
Change-Id: Ibf1c4f85210b9d129424c38d73351df342cea82f