Actually fix UI quirk that leads to Default list being deletable.
Some way or another, the anonymous class that was handling the PopupMenu
listener for Reading List actions was holding on to an incorrect reference
of the ReadingList object. This patch explicitly passes the current
instance of the ReadingList object into the anonymous class, so that the
reference it holds is never ambiguous.

Bug: T185330
Change-Id: I35f420de812f19c1713688178de66deea53b33f4