Update bundled CSS (inc. Cite update)
Updates the bundled CSS.

This includes an update to the ext.cite.style module that will allow the
app to display inline reference links regardless of whether they are
nested in <sup> or <span> elements.

The app will need to keep including the ext.cite.style for as long as MCS is
rewriting sup.mw-ref elements back to spans to support legacy Android
clients.

Bug: T184476
Change-Id: I166b8a1356f91019012064b3eb4959dd3f31f6f0