Fix searching of reading lists.
The previous implementation of reading lists had searching implicitly built into
the db query that fetched the lists.  The current implementation doesn't do this,
so we need to filter the lists at the UI level.

This also improves searching by stripping accents, to allow compatibility
matching of characters with/without accents and diacritics.

Bug: T185495
Change-Id: I1f295d2c03e753fc3cf7e6dc9597d7b16d0e1d7d