Fix display of Default list title in numerous places.
The replacement of the Default list's empty title with the "Saved" string
should be done at the level of the ReadingList class itself.  This will
eliminate the need for any conditional logic anywhere else for determining
whether to display the list title or the default "Saved" title.

This also automatically fixes a number of places where we're incorrectly
showing an empty string instead of the default list title.

Change-Id: Ide69597f4041ded39331695d6dc6dcf5861f0e37