Improve performance of sorting large lists.
The sorting of the pages in a list is done by taking the title of each
page and stripping it of accents, and making it lowercase. These are kind
of expensive-ish operations, and when we have a list of 1000 pages, the
sort operation can actually cause a performance slowdown.

This patch adds a private member to the ReadingListPage class (as well as
the ReadingList class) that lazy-loads and caches the accent-normalized
title, so that during a sort operation it doesn't need to be computed more
than once.

Change-Id: Ic4f6408115f2d9b98fe3e5dae98743bccd51f2d9