Revert "Stop bundling ext.cite.style"
A recent Parsoid markup change[1] is preventing the Android app from
displaying inline reference links.  The change requires either updating
or dropping ext.cite.style modules from the app's style module.

For the benefit of current and prior versions it looks like we'll have to
rewrite the Parsoid markup change back to its original form in MCS in order
to fix this on the service side.[2]

The rewrite means that the app will continue to need the ext.cite.style
module to display reference links in superscript, for as long as this MCS
workaround is required.

[1] I7c83d12dd900682799c124ddae1a8689969d5e8c
[2] I9d531c14180f9099934bff1997ba6d8cbf18458a

Bug: T185427
This reverts commit 07cce4c2902388a1e192d02a436c99d3bb86c3a5.

Change-Id: I829748030a77b6448cb0e44baea129552589fdcf