Explicitly disable caching on service calls that should't be cached.
There was a subtle issue unearthed by Cooltey where certain requests were
incorrectly getting cached, which is the expected behavior in general
(e.g. in the case of article content), but is not correct for things that
shouldn't be cached by definition, such as Reading List collections.

This patch adds an explicit Cache-Control header to all requests that
should never be cached, which includes things like editing, receiving
notifications, and dealing with reading lists.

Change-Id: Ied3a08ca0edf2903d3db8cbb6c8a3f471709f6a1