Follow-up to 'SAVED' button fill state.
This builds on the previous patch by moving the sending of the
ArticleSavedOrDeletedEvent to the lowest possible level (the db layer),
which means that the event will be sent if a page is added or deleted from
*anywhere* in the app, instead of one particular place.

This also adds PageActivity as a consumer of this event, since
PageFragment shows a bookmark icon that needs to be filled/emptied if the
page is added/deleted.

And this also removes the old logic from FeaturedArticleCardView that
explicitly refreshed the Saved button from the Feed, since this is no
longer needed.

Bug: T185498
Change-Id: I34bd36910139e7eff5a99c89ad68cf3ee529e485