Improve injection of testing network interceptor.
This moves our Mock network interceptor into the androidTest source tree,
where it belongs, and makes it injectable into our ConnectionFactory via a
"stub" interceptor that is always added to the Client.

We now inject the interceptor via a custom TestRunner, which is able to
operate on static variables in the app before the activity is launched,
which is better than using the @Setup directive, which is executed only
after the activity is already running.

This also cleans up the interceptor a bit, and removes some #reinvented
logic.

The interceptor now also handles eventlogging messages (returns HTTP 204
status for them), and returns 404 for all other requests, meaning that the
interceptor is now a *complete* handler for all requests coming from the
app.

Change-Id: I8b37054ceca54032497dfcaf1fde30e0e3d1bc6f