Update launch screen graphic.
This updates the launch graphic to be the Wikipedia globe logo, and makes
the launch screen background white.

This also removes an unnecessary duplicate globe resource, as well as
other resolutions of the globe image, especially the xxxhdpi resolution.
Devices are perfectly capable of sizing the provided xxhdpi bitmap to the
appropriate resolution.

Bug: T146395
Change-Id: I85e068d172f24bf00b7c6afbbdcfdd9d845e555a