Apply loading splash screen only to MainActivity.
Since MainActivity is the primary activity that will be launched from the
Home screen, it should really be the only activity that gets the splash
treatment.

Having the splash graphic for all activities is a bit unnecessary, since
the app is fully loaded when most other activities are launched, meaning
that we're asking the system to load the splash bitmap, and then unload it
a few milliseconds later.

This is therefore a slight performance gain for all other activities.

Change-Id: I09f45643f2a8b43d6f5230b13484cbd8708d5288