Integrate with SiteInfo API, with benefits thereof.
This integrates with the `siteinfo` API, which provides us with a lot
of juicy information about the current wiki site.[1]

For now, we will start using the following dynamic information received
from this API:

- The "main page" name of the current site. We currently have the Main
  Page names hard-coded for all wikis, but this will allow us to fetch
  the correct names dynamically, and fall back on the hard-coded names in
  case the API fails, or we're offline.

- The reading list page limit cap. Since this number can vary after our
  rollout of reading list syncinc, it's better if we can get the most
  up-to-date limit dynamically, and fall back on a sane constant.

Check out the other stuff given by the siteinfo API, since I'm sure we can
use more of it in other contexts:

[1] https://en.wikipedia.org/w/api.php?action=query&meta=siteinfo

For now, the app fetches the siteinfo whenever the wiki language is
changed, and then caches the siteinfo statically. This can potentially
create a race condition where the app hasn't yet finished fetching the
siteinfo, so we're forced to get fallback information, but both of the use
cases outlined above are so edge-casey that a race condition doesn't
matter in this case.

Change-Id: Idd14002d7a1ccedc787027d36f974fe6b6d43146