Verify username availability in real time.
This adds logic to verify the availability of a selected user name (when
creating an account) in real time, as the user types the name.

This also consolidates and improves some logic surrounding retrieval of
user info. Specifically it removes the UserIdClient class, since it was
doing nearly identical things to UserExtendedInfoClient.

Bug: T188340
Change-Id: I7863b4f33fedac08acd59901ab40d4329572d6a9