Retire the CSS 'loading' animation from loading.less
The pure-CSS loading animation did good service over the years, but it's
time to put it out to pasture.  The Android app can and should (and does)
indicate that an article is in a loading state with a native ProgressBar.

This is part of a larger project to consolidate WMF mobile client styles
(and another project to retire the MobileApp extension).

Change-Id: I550fa8cd1d059be3da86bcba4761e6e63315df92
Depends-On: Ib3be9f7aef53394866a5956c592d44fb5d810022