Stagger the layout of feed cards in a better sequential order.
Due to the way that the RecyclerView works, when we add lots of cards to
it at one time, it can appear as if the cards are being rendered from the
"bottom up", and can be a bit confusing to the user.

This introduces a short background-thread delay in between inserting of
each card, such that the flow of the feed cards being populated feels much
more natural.

Change-Id: I60ccd591eec32bc20d3dd4026b76c0487e5d75f2