Introduce custom animator for incoming feed cards.
Please, marvel at how insanely difficult Android makes it to properly
customize the animation of RecyclerView items!

The feed cards will now have a brief swipe-up animation when being
inserted for the first time, to create the effect of being "stacked" in
real time against the top.

This also necessitates a slight change to our checkstyle rules, but the
change makes sense anyway.

Bug: T137514
Change-Id: Ib3b18d7ec7609a9516a22344ba90e59e52082a3f