Improve handling of WebView updating exception.
While the system WebView is being updated, our app physically can't
instantiate a WebView object when inflating our PageActivity. The only
thing we can do is fail gracefully and ask the user to try again later.

It turns out that there are a couple of variations of the exception thrown
when the WebView is being updated, some of which we weren't catching. This
patch improves the `catch` block to handle a broader range of WebView-
related exceptions, and should handle all the ones seen in HockeyApp.

Change-Id: If34830166a7be80db8fff22c6616c390894ad80a