Don't mark LicenseActivity as exported.
There's no reason to mark it as exported, since it's only our app that
will ever invoke it.
This activity displays open-source license information for the third-party
libraries that we use, so there's no "security" risk here, but this is for
the sake of correctness.

This also removes an unnecessary intent-filter from AboutActivity, which
allows us to remove the unnecessary `exported` attribute from it, too.

Bug: T189726
Change-Id: I92ab0e9c2416ee6ab012382ddcb20f37e76955d5