Fix bottom margin of article content.
The CSS "loading indicator" was recently removed, but it was actually
contributing to the spacing at the bottom of the article. Since it was
removed, the margin at the bottom is a bit too small.

This adds an explicit height to the bottom div that we place at the end of
the content.

Change-Id: Idf6d39e0202931fb9c22cc2639b837afcde2c364