Respect the new description_source parameter returned by the API.
This starts using the newer `description` parameter that is now available
in the API, in place of the old `pageterms` parameter that will soon be
deprecated.

This also adds the new `descriptionsource` parameter in the MW API, and
`description_source` in RESTBase, which we will use to determine whether
the description is coming from Wikidata or the in-article keyword.

We will now restrict editing descriptions only if the description comes
from the "central" source, which is Wikidata.

Bug: T191337
Change-Id: I20c91ce2927ca3210d309f9d2e8c3868881c39a4