Remove language codes for which the corresponding Wikipedia is closed.
This updates our Python scripts to exclude languages for which the
corresponding Wiki site has been closed, which we can get from the API.

This also uses the REST API to get the total pageviews for each
language-wiki for the previous month, so that we can order the languages
in descending order of usage.

This also removes "None" as a supported language.

This also updates the runtime for the language scripts to specify "python"
(implying Python 3) instead of "python2".

Bug: T156943
Change-Id: I1b4c2a03e1747a83460554be22d9ee0b72dd39a2