Use ESLint for Javascript code.
This switches from jshint to eslint for validating our Javascript code.
This does not yet inherit the eslint configuration from the "wikimedia"
configuration set, since full compliance with those rules would not be a
trivial undertaking. Instead, this basically mimics the same rules we had
under the previous jshint configuration.

Bug: T104108
Change-Id: I6bdc2b6df7ec0a3baef8fb86068f5455cb3c39f1