Use HTML extract from summary
The plain text version has the issue that it discards any styling. This
patch uses the extract_html instead of the plain text extract field
available in the summary for the three areas summaries are used:
* LinkPreviews
* Randomizer
* Featured article

Also updated the unit test fixtures with the latest summary output,
and added a unit test for LinkPreviewContents.

Bug: T191938
Change-Id: I0981a2d174d7972829d7eb82dc2003aeaa22478f