Introduce ABI splits, and reduce APK sizes by 60%.
This introduces the idea of creating multiple APKs for each ABI that we
support. Much of our total APK size is taken up by the binary libraries
from Fresco and Mapbox. The problem is that our final APK contains the
binary libraries for *all four* ABIs that we support (armv7a, arm64, x86,
and x86_64), even though only one will be used on any single device.

Therefore we can now build individual APKs for each of the ABIs, and
upload all of them to the Play Store, which will automatically decide
which one to serve to the device on which the app is being installed. (We
will also continue to produce a "universal" APK for users to download
manually.)

The result is that each individual APK will be nearly 60% smaller (10 MB)
than our current universal APK (23 MB).

Note: Each APK for all the various ABIs will have its own versionCode,
based on this scheme: [ABI code] * 10000 + variant.versionCode.

So, if the current versionCode is 232, the new version code will be 10232
for armv7a, 20232 for arm64, 30232 for x86, and 40232 for x86_64.

https://developer.android.com/studio/build/configure-apk-splits.html

Change-Id: I212dcc9f7eb47199f9f01f55aae8902500fe0d28