Cancelable Retrofit calls will no longer invoke callbacks.
Some of the Retrofit calls that we make are cancelable (using the
.cancel() method).  You would think that if the call is cancelled, it will
never invoke its callback functions onResponse() and onFailure(). However,
this is actually not the case:  the onFailure() callback will still be called,
even if the call was cancelled. Therefore, we need to manually check whether
the call had been cancelled inside our callbacks, and not allow execution
to continue if it's cancelled.

Allowing cancelled calls to continue can lead to unwanted UI behavior,
some of which can be observed when going to Search and typing a few
characters in quick succession.

Change-Id: Icd8109989aac57cdf9a9097a10e72ffe8ac1d4cc