Allow link preview to be scrolled beyond screen height.
Some link previews have quite a bit of text, and sometimes it's enough to
fill the entire screen height, which actually prevents the content to be
scrolled all the way to the bottom. This puts the link preview layout
inside a NestedScrollView, so that it's scrollable all the way, regardless
of screen height constraints.

Change-Id: I75c92de9378fc8d63228c49f827729467d3feae2