Nail down keyboard language detection.
Based on Cooltey's exploratory work, we're now able to leverage the
system's list of installed keyboard languages, and augment our
getAvailableLanguages() function with them.

Since the keyboard languages are added (and merged) after the system
languages are added, it will automatically make the keyboard languages
appear lower in the list than system languages, which is the expectation
for the design of this feature.

Bug: T193047
Change-Id: I5d3518ed0fcc2a560594ddefa127b9a8f01d7cac