Rework/improve handling and persistence of cookies.
Our cookie handling and persistence code was relying on some rather
antiquated logic, and needed a refresh.

This patch moves away from using JavaNetCookieJar, and instead persists
the plain Cookie objects given by OkHttp.

More importantly, it looks like our old code actually wasn't persisting
the *expiry* of the cookies, so it was possible for the app to pass
cookies to the server that have expired. This may have contributed to the
confusion surrounding the `cpPosIndex` cookie.

This stores the cookies in SharedPreferences as before, but serializes
them using Gson with a TypeAdapter, to avoid reflection.

Bug: T190082
Change-Id: Ia92d08966bfd12cf08a317ae4224f7a6884646ab