Correctly format "x years ago" string in OnThisDay.
The current formatting of this text uses a `plurals` resource, which is
not quite correct, since the proper grammar of "x years ago" is very
different for various languages, and can be maddeningly complex with
modulo conditions and special cases.  Currently at least the Russian and
Hebrew localizations of this text come out incorrectly (and are
furthermore impossible to implement correctly with `plurals`).

This patch switches to using the ICU4J functionality that is bundled with
API 24 and higher. (i.e. the updated functionality will only work in API
24 and above.)  If we want to have the same functionality in lower APIs,
we'll need to bundle the ICU library ourselves, but we can visit that
later.

Change-Id: Ifaeda7fa46ca36e45b7d36b7da7ac67f82049d15