Be more fault-tolerant with malformed reading list pages.
Our current sync behavior will completely bail on the sync operation if
one of the reading list articles causes a `no-such-project` error, which
means that if a user manages to add a malformed page, the app will attempt
to sync the page every time, and will never actually complete a successful
sync.

This patch now tolerates `no-such-project` errors from individual pages,
and allows the sync to continue and complete. For the time being, it will
also silently delete the offending page from the list, so that it will not
be attempted to be synced again.

This also hardens the offline-saving logic whereby nonexistent or malformed
pages could have caused errors that would make the service retry saving
the offending page indefinitely.

Bug: T192896
Change-Id: Ie0bc053c28385841b8ae2024ba9bf27499cc4f1e