Fix possible no-such-project errors when saving to reading lists.
There are certain cases where adding a reading list page will save it with
the *mobile* domain name (e.g. en.m.wikipedia.org) instead of the plain
desktop project name (en.wikipedia.org), and the reading list service
seems to be rejecting the mobile versions of the names.

This patch also further improves handling of no-such-project errors
(should they occur for other reasons) by no longer silently deleting the
page, but simply marking it with an invalid remoteID, so that we won't
retry syncing it again, but it will still be available in the user's local
lists.

Change-Id: I4551749a942f10b0a943e7c8a91e7ce68e038c8b