Follow-up: Refine and extend our support for the Referer header.
This moves the idea of a "referrer" into the HistoryEntry class, which is
a prime candidate for storing referrer information, since it already
stores a "source" parameter, which is a pseudo-referrer of sorts.  This
allows the referrer to be passed around along with the HistoryEntry, and
no longer need to be cached in a HashMap.

In addition, I found that external browser apps like Chrome actually
provide a REFERRER parameter in the Intent that they send us, so that we
can extract it from the Intent and use it. This patch introduces the use
of this parameter when opening a link from an external app.

Bug: T192779
Change-Id: I047808c4c3fb6e72fc6bd6bf981762debb6b30cb