Fix/improve differentiation of network and SSL-type exceptions.
In our function that checks whether an exception is a "network" exception,
we check whether the exception and any of its inner causes are instances
of certain known exception types. However, this only checks if the exception
is *precisely* equal to the specified class, whereas we really want to
check if the exception is a *subclass* of the specified class.

This is especially true for SSLException subclasses, of which there are
many, and which currently aren't caught by this function. This causes
erroneous error logs to be sent to HockeyApp, when in fact they can be
safely ignored.

Change-Id: If8021b14d5623375829f80833eda28a12e07642d