Don't send ConfigureFunnel data before receiving Availability response.
When the ConfigureFragment is created, it sends a network request to get
the Feed Availability data.  And only after it receives this data, it
constructs the RecyclerView with the feed content items.

If the user happens to exit the ConfigureFragment *before* the
availability response comes back, the `orderedContentTypes` array will be
empty, and will be incorrectly sent to the Funnel.

This patch makes sure that the Funnel event is sent only if the
availability response has actually returned.

Change-Id: I5209471296229f385be9bf48a9ab791ddf68c002