Prevent possible crash when rearranging languages.
This is a rare crash that happens in API 23, which happens because we're
calling notifyDataSetChanged() while the RecyclerView is still being laid
out.

To reproduce, launch the app on an API 23 device, go to rearrange the
Wikipedia languages, pick up and start sliding one language, then release
and immediately pick up another language (before the first one finishes
animating back to its location).

Change-Id: Id2216918a6b37087fa947b052bd15bc389d69260