Prevent possible crash when showing dialog after saveInstanceState()
This is a pretty obscure crash which I can't seem to reproduce, but seems
to happen when a dialog tries to be shown after saveInstanceState() has
been called.

This patch introduces checks that are nice to have in any case, verifying
that the FragmentManager definitely hasn't yet been destroyed or paused.

https://rink.hockeyapp.net/manage/apps/226649/app_versions/455/crash_reasons/226036110

Change-Id: Ie0f1a9e10cb8c0f69f93fb3ae21fefaf509d65ab