For real, fix audio playback in the app.
This introduces a Javascript transform that takes all the <video> tags in
the document, and sees if they actually need to be turned into <audio>
tags.

When Parsoid generates its output, it actually produces a <video> tag
regardless of whether the underlying media is actually video or audio.
This is a problem because the WebView (and most browsers) apply different
styling to <video> elements versus <audio> elements.

This transform finds <video> elements that have a source with an "*.ogg"
extension, which implies that it's an audio file, and performs the
following changes:
- Change the tag from <video> to <audio>
- Remove its "poster" attribute.

Bug: T193478
Bug: T198504
Change-Id: I63cb9a299c3cc3b7258f79de97209a5a883563d6