Move page title/description/etc to WebView.
This moves the logic to display the article title, description, main edit
pencil, and audio pronunciation button to the WebView, and no longer
display them as native Views.

This provides a dramatic simplification for our code, and can eventually
be moved into the page-library, or even the content service.

The only thing remaining that is handled natively at the top of the
article is the lead image. This may eventually open the door to using a
proper CollapsingToolbarLayout in the article reading experience.

Bug: T150459
Bug: T150424
Change-Id: Iff1162f3936872ac8a4d77da0680a994b17a9272