Optimization: make list of Tabs less-often serialized/deserialized.
This is one of several revelatory optimizations that has come out of the
building of the two new prototypes.

There are several places in the app that deal with the list of Tabs that
belong to the current user. However, every time the list of tabs is
queried, it gets read from SharedPreferences and deserialized with Gson,
which is not a trivial operation.

This patch makes the list of tabs get deserialized *once*, when the app
itself initializes, and then gets serialized only when PageFragment is
paused.

No visual or behavioral changes are intended.

Change-Id: I550e58c48d3326f4221083f8b014ab8446b2ee81