Optimization: use DiffResult for notifying changes in RecyclerView.
For a long time I've observed a bit of jankiness when switching back and
forth to the ReadingLists fragment, which means that something was
happening on the main thread that was slowing down the UI.

After a bit of investigating, it looks like we re-query the collection of
reading lists whenever you switch to the ReadingLists fragment, and when
the query is complete, we invalidate the RecyclerView and cause it to
recreate and re-bind all of the list items. This is a problem, and a
common cause of performance degradation.

The proper way of handing this kind of situation is to use DiffUtil to
produce a DiffResult and feed it to the RecyclerView adapter.

With this patch, the interaction with the ReadingLists fragment becomes
smooth as butter.

And one other thing: the main ViewPager will now pre-load up to 2
offscreen fragments, meaning that it will pre-load ReadingLists and
History when the user looks at the Feed, making scrolling between them
even smoother.

Change-Id: I90c63170337d1fea1fcd93840a8419c1af06eda2