Notifications: expand and solidify poll broadcast receiver.
This fleshes out our nascent BroadcastReceiver that does the actual work
of receiving poll events from the system and querying the API for new
notifications in the background, and actually queueing the notification to
be shown in the system status bar.

This also builds a little further on the notification client structures.

Change-Id: Ia04a4d5a7124d6f646b35655cd1737bb7f8d5bcf