Hygiene: escape apostrophes when generating language lists.
In preparation for working with build-tools 28.x.x, they are more strict
with apostrophes in XML lists, and will no longer work with our language
list because at least one language contains an apostrophe in its name.

This patch updates the script to properly escape apostrophes, and performs
a run of the script which refreshes the language lists.

Change-Id: I2bf098c737a1d929062ffe55009a972b734feb1a