Huge optimization: rework Retrofit network layer.
This patch reworks our usage of Retrofit to match the way it was actually
*meant* to be used. This introduces a `Service` class that represents the
entirety of the API surface of our backend, instead of a fragmented
service model where each specific class was implementing its own "Service"
layer. This was creating significant memory and performance bottlenecks.

This also introduces a cache of Retrofit Service objects, since we
communicate with more than one domain at a time, whereas each Retrofit
object is always locked into a single domain. The cache is limited to
eight (8) such objects, which should be sufficient for the great majority
of users, even with multiple languages selected.

Change-Id: I29086b571220d2bdda168c22d1a63059cb7a827b