Rx: Correction - fix usage of doFinally clauses.
In some of our Observable subscriptions, we want to perform an action
regardless of whether the observable was successful or not, sort of like a
"finally" block.  I thought that this could be done with the onComplete()
parameter in the subscribe() call, but this is actually false: the
onComplete() function does *not* get called if an error occurs, because
technically it means that the observable did not "complete" successfully.

Instead, the Observable offers a doFinally() clause where we can put a
lambda that will be executed regardless of the result of the Observable.

Change-Id: I6fc3560818121c51c2d6d2d8d352df46903c2fb8