Fix crash due to recent internal WebView changes.
Some recent updates to the Android System WebView seem to have caused a
regression in the way it handles passing messages to the Java layer: it
seems to be placing an upper limit on the size of messages that can be
sent to the onJsPrompt() callback. This limit appears to be 10240 bytes,
which is OK for most articles, but certain articles contain "issues"
elements that exceed this size, and are therefore causing problems.

I have now resolved the issue properly:

Messages from the WebView shall now be passed using a @JavascriptInterface
method, which does not have a size limitation on its parameters. I'm
actually not sure why we didn't do this from the beginning.  This way, we
also don't need to URL-encode the message in the WebView and then decode
it in Java, making it a bit more efficient.

Here is the associated Chromium issue, for reference:
https://bugs.chromium.org/p/chromium/issues/detail?id=884100

Bug: T205305
Change-Id: I97a5298ac668dfd9705ee7d39720a4b25fca6d81