Fix testability of RestBase service calls, at the expense of a bit of performance.
This breaks up our Service class into two components: Service and
RestService. This is to make the Rest service calls testable by specifying
a custom (non-production) base URL (as we've done previously using
developer preferences).

The problem is that for non-production testing of the Rest service,
different language Wikis are specified in the *path* of the URL, instead
of in the subdomain. This means we can no longer use the same Retrofit
object for all service calls.

This unfortunately sacrifices a bit of performance for the sole purpose of
being able to test the service functionality. Ideally we should work
towards a test system where we can specify the same base URL for both MW
and Rest calls, and recombine the Service into a single class, like
before.

Change-Id: I57cba0656056bbc9ff72a63ebc0634077c8cdd95