Simplify/enhance error processing in MediaWiki API responses.
Previously, if a MediaWiki response contained an error, then every
individual consumer would need to manually check for the presence of an
error.  With this patch, the error will automatically be detected and
thrown as an exception, to be caught at the end of the Rx stream.

This is done by leveraging our functionality of "post-processing" Gson
objects, whereby we can simply take the MwResponse class and post-process
it to check whether it contains an error.

Change-Id: I89bf72e898e2d0c2ed028d0f7ebfb7f732c6f5a2