W-icon behavior with a bit of fudge.
This patch nails down the behavior of the W icon, by introducing a sort of
workaround when navigating back from PageActivity to MainActivity. Here's
the problem:

When we tap the W icon, we want to launch MainActivity with the CLEAR_TOP
flag set, so that the system automatically goes back to MainActivity
regardless of any other activities that are in the current backstack. The
thing is that when we do this, the expected shared-element transition does
not work. If MainActivity already exists in the backstack, it's simply
brought to the front and given an updated Intent, and *not* given a new
transition animation.

This patch works around this issue by using an explicit custom animation
that "simulates" the minimization of PageActivity to the spot where it's
expected to exist in the floating-queue component in MainActivity.

The animation is not "perfect", but it's actually smoother than a shared
element transition, and works well enough to direct the user's gaze to the
floating queue, which is the overall goal of our new navigation model.

Bug: T204544
Change-Id: I268903a5d711286ce00405ee1f4f26d99f2949fd